[fix] 재로그인 시 로그아웃 안 되는 오류 수정 로그아웃 hook 생성
